requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
fitz==0.0.1.dev2
frozenlist==1.7.0
fsspec==2025.7.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
griffe==1.12.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.1
llama-cloud==0.1.37
llama-cloud-services==0.6.60
llama-index-core==0.13.2
llama-index-instrumentation==0.4.0
llama-index-workflows==1.3.0
llama-parse==0.6.60
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psycopg2==2.9.10
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.3
yarl==1.20.1
